backend/requirements.txt
Django==5.1.4
django-cors-headers==3.14.0
celery==5.3.0
redis==4.5.0
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.7.0